accounts/views.py
from rest_framework import generics
from .models import User
from accounts.serializers import UserRegisterSerializer

class RegisterView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = UserRegisterSerializer