requirements.txt
Django==3.2.8
django-filter==21.1
django-password-validators==1.6.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0